feat: 🎸 Add intro and reset password page, and refactor pages
Added the intro page inside the product page because it is linked to
that. It should only be displayed if the user is not logged in. added
reset password and put it in a folder called issue-pages, that is meant
for all issues like: reset password, error page for design not found or
other problems and page-not-found. I also created a folder for
terms-and-conditions that should contain all T&C pages for sales and use
of page. I expect we will have more of these pages in the future.

✅ Closes: #40